refactor: incorporate code reviews
