+++ src/components/atoms/DateText/styles.module.scss
@@ -4,6 +4,7 @@
 .date_text {
   font-size: 0.875rem;
   line-height: 1.5;
+  font-weight: 700;
   @include mixin.pc {
     font-size: 1.5rem;
   }
