+++ src/components/atoms/Tag/styles.module.scss
@@ -16,15 +16,3 @@
   padding: 0px 4px;
   color: variables.$white;
 }
-
-.orange {
-  background-color: variables.$orange;
-}
-
-.red {
-  background-color: variables.$red;
-}
-
-.yellow {
-  background-color: yellow;
-}
