+++ src/components/molecules/WideCard/index.tsx
@@ -1,21 +1,14 @@
-import { Dates } from 'src/components/atoms/Dates';
+import { DateText } from 'src/components/atoms/DateText';
 import { CardText } from 'src/components/atoms/CardText';
-import { Tags } from 'src/components/atoms/Tags';
+import { Tag, TagProps } from 'src/components/atoms/Tag';
 import styles from './styles.module.scss';
-
-type TagTypes = {
-  id: string;
-  name: string;
-  color: string;
-  tagGroup: string;
-};
 
 interface WideCardProps {
   id: string;
   title: string;
   detail: string;
   announcedAt: string;
-  tags: TagTypes[];
+  tags: TagProps[];
 }
 
 export const WideCard: React.FC<WideCardProps> = ({
@@ -25,9 +18,9 @@ export const WideCard: React.FC<WideCardProps> = ({
 }) => (
   <div className={styles.card_container}>
     <div className={styles.card_title_container}>
-      <Dates date={announcedAt} />
+      <DateText date={announcedAt} />
       {tags.map((tag) => (
-        <Tags
+        <Tag
           key={tag.id}
           id={tag.id}
           name={tag.name}
